pyproject.toml
[tool.poetry]
name = "product-review-intelligence"
version = "0.1.0"
description = "Multi-Agent Product Analytics with LLM Intelligence and Quality Monitoring"
authors = ["Bhavya Likhitha Bukka"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
datasets = "^3.0"
pandas = "^2.2"
pyarrow = "^18.0"
python-dotenv = "^1.0"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
loguru = "^0.7"
spacy = "^3.8.14"
matplotlib = "^3.10.8"
seaborn = "^0.13.2"
wordcloud = "^1.9.6"
chromadb-client = "^1.5.5"
snowflake-sqlalchemy = "^1.9.0"
snowflake-connector-python = "^4.4.0"
groq = "^1.1.2"
torch = "^2.11.0"
transformers = "^5.4.0"
mlflow = "^3.10.1"
scikit-learn = "^1.8.0"
joblib = "^1.5.3"
sentence-transformers = "^5.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^24.0"
isort = "^5.13"
flake8 = "^7.0"
pre-commit = "^3.6"
ipykernel = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
